+++ components/HomeScreen/Styles.js
@@ -1,0 +1,36 @@
+import { StyleSheet } from "react-native";
+import {heightPercentageToDP as hp, widthPercentageToDP as wp} from "react-native-responsive-screen";
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#E5FFFF',
+    alignItems: 'center',
+  },
+  titleContainer: {
+    marginTop: wp('35%'),
+
+  },
+  title: {
+    fontSize: wp('8%'),
+
+  },
+  buttonContainer: {
+    flex: .8,
+    justifyContent: 'center',
+  },
+  button: {
+    height: hp('7%'),
+    width: wp('35%'),
+    borderRadius: wp('10%'),
+    margin: 5,
+    backgroundColor: 'black',
+    alignItems: 'center',
+    justifyContent: 'center'
+  },
+  buttonText: {
+    color: 'white',
+  },
+});
+
+export default styles;
